backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
email-validator>=2.1.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
soundfile>=0.12.0
numpy>=1.26.0,<2.0.0
onnxruntime>=1.20.0
ctranslate2>=4.4.0
sentencepiece>=0.2.0
insanely-fast-whisper>=0.0.15
sherpa-onnx>=1.12.0
python-multipart>=0.0.9
websockets>=12.0
starlette>=0.38.0
loguru>=0.7.0
watchfiles>=0.24.0
librosa>=0.10.0
pytest>=8.3.0
httpx>=0.27.0
silero-vad
webrtcvad>=2.0.10
bcrypt>=3.2,<4  # pin for passlib 1.7.x compatibility
passlib[bcrypt]>=1.7.4
Authlib>=1.3.0
python-dateutil>=2.9.0
itsdangerous>=2.2.0
